Change the inline function to be observed: swap()
On the previous commit 607327c2, the function `bbsort()` only be called
one by `main()`. In order to observe the case that an inline function be
called more than once, function specifiers of `bbsort()` and `swap()`
are exchanged. The static object which value to be observed is moved
into `swap()` also.

The `swap()` function is inlined by adding the `always_inline`
attribute which observed by assembly of *bbsort_inline.c*.
The function is inlined to the caller is substituting the line to call
it by its context, the line is in a nested loop, so, I expect that the
calls to it may be duplicated, but it don't.

So, the value of the static object is the same as the object be used as
an automatic variable.

Note: (TO-BE-TESTED) the value may be changed if we unroll the nested
loop and execute iterations of the loop parallelly.