Output value of variables in recipe to stdout
The feature for debugging is copied from the Makefile in lab0-c
of linux2022.

A requirement is that variables in this Makefile are update by
extended Makefile, this feature is added for checking
values of the variables are updated correctly.

A option `VERBOSE` to turn on/off this feature, default is off,
set `1` to turn it on.